Update circleci script to build on non development branches correctly
The make update assumes that develop builds always with the most recent
submodule updates. This is not needed for tagged builds and for non develop builds.
For theses builds, a git submodule update --recursive should be sufficient as it is
setting up the intended dependencies. The git submodule init command can be merged with
update by using --init option.

See also #459, #463
Closes #467

Signed-off-by: Marcel Wagner <wagmarcel@web.de>