Fix the deploy docker action for the nightly
Github Actions overrides the home and working directory parameters
of the Dockerfile, therefore we need to set them correctly for both
kubectl and helm, since they try to search the home directories for
their config files.

Since the namespace of the latest.oisp.info is different than the
default, we add the NAMESPACE secret.

Follow-up to: #556

Signed-off-by: Meric <meric.feyzullahoglu@gmail.com>