Recreate registry on every workflow run
Since we run our workflows on a self hosted environment, we need to
take it into account that if we continue to push new images to our
local registry each run it will fill up and clog the machine. This
commit adds a new makefile command ```make recreate-registry```
which deletes and recreates the local registry and results in a clean
registry each run.

Signed-off-by: Meric Feyzullahoglu <meric.feyzullahoglu@gmail.com>